# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HARROWGATE Visa Consultancy API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Storage
    description: Object storage upload and serving endpoints
  - name: clients
    description: Client management
  - name: applications
    description: Application management
  - name: documents
    description: Document management
  - name: stats
    description: Dashboard statistics
paths:
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content
          content:
            '*/*':
              schema:
                type: string
                format: binary
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve a private object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Not found

  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  /clients/{clientId}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List all applications
      parameters:
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create a new application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  /applications/{applicationId}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get an application by ID
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update an application
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  /applications/{applicationId}/extract:
    post:
      operationId: extractApplicationData
      tags: [applications]
      summary: Use AI to extract data from uploaded documents and fill the application form
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Extraction result with updated form data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractionResult"
        "404":
          description: Not found

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for an application
      parameters:
        - name: applicationId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Register an uploaded document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{documentId}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /stats:
    get:
      operationId: getDashboardStats
      tags: [stats]
      summary: Get dashboard summary statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Client:
      type: object
      required: [id, name, email, nationality, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        nationality:
          type: string
        countryOfOrigin:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        passportNumber:
          type: ["string", "null"]
        createdAt:
          type: string

    ClientInput:
      type: object
      required: [name, email, nationality]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        nationality:
          type: string
        countryOfOrigin:
          type: string
        dateOfBirth:
          type: string
        passportNumber:
          type: string

    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        nationality:
          type: string
        countryOfOrigin:
          type: string
        dateOfBirth:
          type: string
        passportNumber:
          type: string

    Application:
      type: object
      required: [id, clientId, status, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, documents_uploaded, ai_processing, ai_processed, ready_to_print, submitted]
        targetProgram:
          type: ["string", "null"]
        targetUniversity:
          type: ["string", "null"]
        applicationType:
          type: ["string", "null"]
          enum: [student_visa, pr_pathway, university_admission, null]
        formData:
          $ref: "#/components/schemas/FormData"
        notes:
          type: ["string", "null"]
        documentCount:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    ApplicationInput:
      type: object
      required: [clientId]
      properties:
        clientId:
          type: integer
        targetProgram:
          type: string
        targetUniversity:
          type: string
        applicationType:
          type: string
        notes:
          type: string

    ApplicationUpdate:
      type: object
      properties:
        status:
          type: string
        targetProgram:
          type: string
        targetUniversity:
          type: string
        applicationType:
          type: string
        formData:
          $ref: "#/components/schemas/FormData"
        notes:
          type: string

    FormData:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        passportNumber:
          type: ["string", "null"]
        passportExpiry:
          type: ["string", "null"]
        homeAddress:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        highestDegree:
          type: ["string", "null"]
        institution:
          type: ["string", "null"]
        graduationYear:
          type: ["string", "null"]
        gpa:
          type: ["string", "null"]
        languageTestScore:
          type: ["string", "null"]
        languageTestType:
          type: ["string", "null"]
        intendedProgram:
          type: ["string", "null"]
        intendedUniversity:
          type: ["string", "null"]
        fundingSource:
          type: ["string", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]

    Document:
      type: object
      required: [id, applicationId, documentType, fileName, uploadedAt]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        documentType:
          type: string
          enum: [passport, academic_transcript, degree_certificate, language_certificate, bank_statement, recommendation_letter, personal_statement, photo, other]
        fileName:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: ["integer", "null"]
        mimeType:
          type: ["string", "null"]
        extractedText:
          type: ["string", "null"]
        uploadedAt:
          type: string

    DocumentInput:
      type: object
      required: [applicationId, documentType, fileName, fileUrl]
      properties:
        applicationId:
          type: integer
        documentType:
          type: string
        fileName:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: integer
        mimeType:
          type: string

    ExtractionResult:
      type: object
      required: [success, formData]
      properties:
        success:
          type: boolean
        formData:
          $ref: "#/components/schemas/FormData"
        extractedFields:
          type: array
          items:
            type: string
        message:
          type: string

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    DashboardStats:
      type: object
      required: [totalClients, totalApplications, pendingReview, readyToPrint, submitted]
      properties:
        totalClients:
          type: integer
        totalApplications:
          type: integer
        pendingReview:
          type: integer
        readyToPrint:
          type: integer
        submitted:
          type: integer
        recentApplications:
          type: array
          items:
            $ref: "#/components/schemas/Application"